pyproject.toml
[tool.poetry]
name = "jdna"
version = "0.1.7"
description = ""
authors = ["Justin Vrana <justin.vrana@gmail.com>"]
homepage = 'https://github.com/jvrana/jdna'
repository = 'https://github.com/jvrana/jdna'
documentation = 'https://jvrana.github.io/jdna/'
readme = 'README.md'
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Development Status :: 4 - Beta"
]

[tool.poetry.dependencies]
python = "^3.5"
biopython = "^1.73"
networkx = "^2.3"
primer3-py = "^0.6.0"
colorama = "^0.4.1"
webcolors = "^1.9"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", python = "^3.6", allows-prereleases = true}
pytest = "^5.0"
pre-commit = "^1.17"
toml = "^0.10.0"
pytest-benchmark = "^3.2"
guzzle_sphinx_theme = "^0.7.11"
flake8 = "^3.7"
docformatter = "^1.3"
tox = "^3.14"
tox-conda = "^0.2.1"

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
